fixes#7(back-end)

+++ server/src/controllers/categories.controller.ts
@@ -1,12 +1,26 @@
 import { Request, Response, NextFunction } from 'express'
+import * as yup from 'yup'
 import ApiError from '../errors/ApiError'
 import categoriesService from '../services/categories.service'
 import { ReturnType } from './types/return.type'
 import Category from '../db/models/category.model'
 import Objection from 'objection'
 import { ICategoryBody } from '../services/types/categories.type'
+import { addCategorySchema, updCategorySchema } from './types/schemas'
 
 export default class CategoriesController {
+  static async getCategoryById(req: Request, res: Response, next: NextFunction): ReturnType<Objection.Page<Category>> {
+    const { id } = req.params
+
+    const category = await categoriesService.getCategoryById(+id)
+
+    if (!category) {
+      return next(ApiError.badRequest(`Fetching category error`))
+    }
+
+    return res.json(category)
+  }
+
   static async getCategories(req: Request, res: Response, next: NextFunction): ReturnType<Objection.Page<Category>> {
     const categoryId = +req.query.id || null
     const categoryName = req.query.name || null
@@ -31,12 +45,14 @@ export default class CategoriesController {
     const categoryName: string = req.body.name
     const parent: number = req.body.parent ? +req.body.parent : 0
 
-    if (!categoryName) {
-      return next(ApiError.internal('Please, type the category name'))
-    }
+    await addCategorySchema.validate({ categoryName, parent })
 
-    if (parent !== 0 && !parent) {
-      return next(ApiError.internal('Please, type the category parent'))
+    const productCategory = await Category.query().findOne({
+      name: categoryName,
+    })
+
+    if (productCategory) {
+      return next(ApiError.internal('Category with this name already exists'))
     }
 
     const category = await categoriesService.addCategory({
@@ -59,6 +75,14 @@ export default class CategoriesController {
       return next(ApiError.badRequest('Please, type the category id'))
     }
 
+    const oldCategory = await Category.query().select().findById(+categoryId)
+
+    if (!oldCategory) {
+      return next(ApiError.internal("Can't find this category"))
+    }
+
+    await updCategorySchema.validate(changingValues)
+
     const category = await categoriesService.updateCategory(+categoryId, {
       name: changingValues.name ? changingValues.name : null,
       parent: changingValues.parent ? +changingValues.parent : null,
@@ -78,16 +102,18 @@ export default class CategoriesController {
       return next(ApiError.internal('Please, type the category id'))
     }
 
+    const oldCategory = await Category.query().select().findById(+categoryId)
+
+    if (!oldCategory) {
+      return next(ApiError.internal("Can't find this category"))
+    }
+
     const result = await categoriesService.deleteCategory(+categoryId)
 
     if (!result) {
       return next(ApiError.badRequest(`Deletion category error`))
-    }
-
-    if (typeof result == 'number') {
-      return res.json({ message: `Successfully deleted ${result} categories` })
     } else {
-      return res.json(result)
+      return res.json({ message: `Successfully deleted ${result} categories` })
     }
   }
 }
